refactor: turn types into a subpackage
